[ONNX] Moved uninitialized_optional test cases into runtime test and parameterized left test cases (#80145)
- moved test_uninitialized_optional into runtime test, because **OptionalHasElement** node now is supported on onnxruntime. test_optional_input will be moved later when tracing/None output paradox is fixed.
- changed _TestONNXRuntime to inherited by common_utils.Testcase as it's more rich on error message
- modified the left itertools implemented test cases to using parametrized decorator
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80145
Approved by: https://github.com/justinchuby, https://github.com/BowenBao
